requirements.txt
# Core dependencies
opencv-python-headless>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
Flask>=2.0.0
Werkzeug>=2.0.0
scikit-image>=0.18.0
scipy>=1.7.0

# Production dependencies
Flask-Limiter>=2.0.0
Flask-Caching>=1.10.0
gunicorn>=20.1.0

# Security and monitoring
python-dotenv>=0.19.0
psutil>=5.8.0

# Optional: Database support (uncomment if needed)
# SQLAlchemy>=1.4.0
# Flask-SQLAlchemy>=2.5.0
# alembic>=1.7.0

# Optional: Redis for caching and rate limiting (uncomment if needed)
# redis>=4.0.0
# Flask-Limiter[redis]>=2.0.0

# Optional: Monitoring and logging (uncomment if needed)
# prometheus-client>=0.12.0
# structlog>=21.1.0
